pyproject.toml
[project]
name = "main-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.17.0",
    "cohere-haystack>=5.3.0",
    "faiss-cpu>=1.12.0",
    "haystack-ai>=2.17.1",
    "openai>=1.101.0",
    "opencv-python>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.1.0",
    "spacy>=3.8.7",
    "streamlit>=1.48.1",
    "striprtf>=0.0.29",
    "torch>=2.8.0",
    "transformers>=4.56.0",
]
